# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Oblivious API
  description: |
    Oblivious 是一个面向工作区的 AI 应用平台，提供对话、知识库、任务编排等能力。
    本文档描述了所有可用的 REST API 端点。
  version: "1.0.0"
  contact:
    name: Oblivious Team
servers:
  - url: http://localhost:8080
    description: 本地开发环境
  - url: https://api.oblivious.dev
    description: 生产环境
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: 登录后通过 Set-Cookie 设置的会话令牌
  schemas:
    Envelope:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        data:
          description: 成功时返回的业务数据
        error:
          $ref: "#/components/schemas/ErrorPayload"
    ErrorPayload:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: invalid_request
        message:
          type: string
          example: "invalid json body"
...
    KnowledgeBase:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        documentCount:
          type: integer
    CreateKnowledgeBaseRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "产品文档"
    Document:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
    KnowledgeChunkMetadata:
      type: object
      properties:
        documentVersion:
          type: string
        pageNumber:
          type: integer
        sourceUrl:
          type: string
        startRune:
          type: integer
        endRune:
          type: integer
    KnowledgeDocumentChunk:
      type: object
      properties:
        chunkId:
          type: string
        chunkIndex:
          type: integer
        content:
          type: string
        documentVersion:
          type: string
        charCount:
          type: integer
        estimatedTokenCount:
          type: integer
        metadata:
          $ref: "#/components/schemas/KnowledgeChunkMetadata"
    CreateDocumentRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        content:
          type: string
    UpdateKnowledgeDocumentChunkRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
    SplitKnowledgeDocumentChunkRequest:
      type: object
      required: [splitAt]
      properties:
        splitAt:
          type: integer
          minimum: 1
    MergeKnowledgeDocumentChunksRequest:
      type: object
      required: [direction]
      properties:
        direction:
          type: string
          enum: [previous, next]
    RetrieveKnowledgeRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          example: "如何配置模型"
...
  responses:
    Unauthorized:
      description: 未认证或会话已过期
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
          example:
            ok: false
            data: null
            error:
              code: unauthorized
              message: "authentication required"
...
    BadRequest:
      description: 请求参数无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
          example:
            ok: false
            data: null
            error:
              code: invalid_request
              message: "invalid json body"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
          example:
            ok: false
            data: null
            error:
              code: not_found
              message: "resource not found"
...
paths:
...
  /api/v1/app/knowledge-bases:
    get:
      summary: 获取知识库列表
      operationId: listKnowledgeBases
      tags: [Knowledge]
      responses:
        "200":
          description: 知识库列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/KnowledgeBase"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: 创建知识库
      operationId: createKnowledgeBase
      tags: [Knowledge]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKnowledgeBaseRequest"
      responses:
        "200":
          description: 新创建的知识库
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/KnowledgeBase"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/app/knowledge-bases/{knowledgeBaseId}:
    get:
      summary: 获取知识库详情
      operationId: getKnowledgeBase
      tags: [Knowledge]
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 知识库详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/KnowledgeBase"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      summary: 更新知识库
      operationId: updateKnowledgeBase
      tags: [Knowledge]
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKnowledgeBaseRequest"
      responses:
        "200":
          description: 更新后的知识库
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/KnowledgeBase"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      summary: 删除知识库
      operationId: deleteKnowledgeBase
      tags: [Knowledge]
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: 删除成功
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/app/knowledge-bases/{knowledgeBaseId}/documents:
    get:
      summary: 获取知识库文档列表
      operationId: listKnowledgeDocuments
      tags: [Knowledge]
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 文档列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Document"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: 创建知识库文档
      operationId: createKnowledgeDocument
      tags: [Knowledge]
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentRequest"
      responses:
        "200":
          description: 新创建的文档
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Document"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/app/knowledge-bases/{knowledgeBaseId}/documents/{documentId}:
    put:
      summary: 更新知识库文档
      operationId: updateKnowledgeDocument
      tags: [Knowledge]
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: string
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentRequest"
      responses:
        "200":
          description: 更新后的文档
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Document"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      summary: 删除知识库文档
      operationId: deleteKnowledgeDocument
      tags: [Knowledge]
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: string
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: 删除成功
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/app/knowledge-bases/{knowledgeBaseId}/documents/{documentId}/chunks:
    get:
      summary: 获取知识库文档分块
      operationId: listKnowledgeDocumentChunks
      tags: [Knowledge]
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: string
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 文档分块列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/KnowledgeDocumentChunk"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/app/knowledge-bases/{knowledgeBaseId}/documents/{documentId}/chunks/{chunkId}:
    put:
      summary: 更新知识库文档分块
      operationId: updateKnowledgeDocumentChunk
      tags: [Knowledge]
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: string
        - name: documentId
          in: path
          required: true
          schema:
            type: string
        - name: chunkId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKnowledgeDocumentChunkRequest"
      responses:
        "200":
          description: 更新后的文档分块
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/KnowledgeDocumentChunk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/app/knowledge-bases/{knowledgeBaseId}/documents/{documentId}/chunks/{chunkId}/split:
    post:
      summary: 拆分知识库文档分块
      operationId: splitKnowledgeDocumentChunk
      tags: [Knowledge]
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: string
        - name: documentId
          in: path
          required: true
          schema:
            type: string
        - name: chunkId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SplitKnowledgeDocumentChunkRequest"
      responses:
        "200":
          description: 拆分后的当前文档分块列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/KnowledgeDocumentChunk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/app/knowledge-bases/{knowledgeBaseId}/documents/{documentId}/chunks/{chunkId}/merge:
    post:
      summary: 合并知识库文档分块
      operationId: mergeKnowledgeDocumentChunks
      tags: [Knowledge]
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: string
        - name: documentId
          in: path
          required: true
          schema:
            type: string
        - name: chunkId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MergeKnowledgeDocumentChunksRequest"
      responses:
        "200":
          description: 合并后的当前文档分块列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/KnowledgeDocumentChunk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/app/knowledge-bases/{knowledgeBaseId}/retrieve:
    post:
      summary: 知识库检索
      operationId: retrieveKnowledge
      tags: [Knowledge]
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetrieveKnowledgeRequest"
      responses:
        "200":
          description: 检索结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ---- 任务 (Task) ----
...